refactor: rename app.documents_pkg → app.judgments_pkg
Directory rename only. Local alias `documents_router` in server.py and the
URL prefix `/documents` remain unchanged — those move in later phases.
