+++ backend/app/recommendations.py
@@ -16,7 +16,7 @@ from loguru import logger
 from pydantic import BaseModel, Field
 
 from app.core.supabase import get_supabase_client
-from app.documents_pkg import generate_embedding
+from app.judgments_pkg import generate_embedding
 from app.models import validate_id_format
 
 router = APIRouter(prefix="/recommendations", tags=["recommendations"])
