+++ backend/app/server.py
@@ -41,7 +41,7 @@ from app.clustering import router as clustering_router
 from app.collections import router as collections_router
 from app.dashboard import router as dashboard_router
 from app.deduplication import router as deduplication_router
-from app.documents_pkg import router as documents_router
+from app.judgments_pkg import router as documents_router
 from app.embeddings_api import router as embeddings_router
 from app.evaluations import router as evaluations_router
 from app.example_questions import router as example_questions_router
